backend/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
gunicorn==21.2.0
python-dotenv==1.0.0
openai>=1.3.0
websockets==12.0
python-multipart==0.0.6
httptools==0.6.1
pydantic==2.5.2
pydantic-settings==2.1.0
python-jose==3.3.0

# HTTP and async dependencies
aiohttp==3.9.0
httpx==0.25.1
async-timeout==4.0.3
aiofiles==23.2.0

# Utilities
ujson==5.8.0
orjson==3.9.10
python-dateutil==2.8.2
typing-extensions>=4.8.0

# WebSocket dependencies
starlette==0.27.0
uvicorn[standard]==0.24.0

# Production dependencies
gunicorn==21.2.0
uvloop==0.19.0; sys_platform != 'win32'
httptools==0.6.1

# Logging and monitoring
structlog==23.2.0
prometheus-client==0.18.0

# Security
cryptography==41.0.5
passlib==1.7.4
bcrypt==4.0.1

# Azure integration
azure-identity==1.14.1
azure-search-documents==11.4.0

# Additional tools
PyMuPDF==1.22.5  # fitz
chardet==5.1.0
youtube-transcript-api==0.6.0
requests==2.31.0
python-docx==0.8.11
python-pptx==0.6.21
beautifulsoup4==4.12.2
pandas==2.0.3